added client reconnection for all cases
If a Client disconnects he will reconnect to the server.
Depending on the state of the client different Events occur.
If the Player disconnected in any state except he was already in a running game he will just get the notification if the connection is established.
If he joined a game and its already running he will get a different request type (RECONNECT_TO_GAME).