small changes to cheating functionality
When someone submits a cheat the money gets updated and now it gets handled the same way as all update money actions. And if someone detects a cheating Player now two seperate calls gets send to the Players. One to update the money of the cheating player and one to notify a cheater has been exposed.